Backport e3695a1c8ff6443ede36c4f3bfe80b9c7fb0ed22: update all events to macros
